backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
boto3==1.35.0
mangum==0.17.0
pydantic==2.8.0
requests==2.32.0
